Demonstrate image distortion due to CSS styling
When hard values for image height/width are added
in CSS styles, they may distort the original
image. This can be avoided by setting a hard value
for one of the dimensions (either height or width)
and setting the dimension to "auto," which
automatically makes the image proportional with
new dimensions.